requirements.txt
pytorch
pandas
matplotlib
seaborn
h5py
scikit-learn
scikit-image
tqdm
packaging
